Stock, Scatterer, EVE working fine
Fixed the stock rim to follow the correct sun (the nearest).

Scatterer compatibility done (the mod is looking for materials of both original and 'synchronize' them to load / unload scatterer as wanted.

EVE is working by itself; However the tilt settings has to be done in the EVE config files (the Tilt addon is not touching any of it)